parser-json-sarif: separate module for SarifTreeDecoder
